+++ src/fiber.rs
@@ -1,9 +1,9 @@
-use super::bamranges::*;
-use super::basemods::BaseMods;
 use super::subcommands::center::CenterPosition;
 use super::subcommands::center::CenteredFiberData;
 use super::utils::FiberFilters;
 use super::*;
+use crate::utils::bamranges::*;
+use crate::utils::basemods::BaseMods;
 use crate::utils::bio_io::*;
 use rayon::prelude::*;
 use rust_htslib::bam::Read;
