+++ src/utils.rs
@@ -5,6 +5,9 @@ use rust_htslib::bam;
 use rust_htslib::bam::Read;
 use std::fmt::Debug;
 
+pub mod bamlift;
+pub mod bamranges;
+pub mod basemods;
 pub mod bio_io;
 pub mod fire;
 pub mod nucleosome;
