+++ src/utils/basemods.rs
@@ -1,4 +1,4 @@
-use super::bamranges::*;
+use crate::utils::bamranges::*;
 use crate::utils::bio_io::*;
 use bio::alphabets::dna::revcomp;
 use lazy_static::lazy_static;
